+++ CyRK/cy/common.hpp
@@ -39,7 +39,11 @@ static const double DYNAMIC_GROWTH_RATE = 1.618;
 static constexpr double SIZE_MAX_DBL = 0.99 * SIZE_MAX;
 
 
-typedef void (*DiffeqFuncType)(double*, double, double*, const void*);
+
+typedef void (*PreEvalFunc)(void*, double, double*, const void*);
+
+typedef void (*DiffeqFuncType)(double*, double, double*, const void*, PreEvalFunc);
+
 
 
 struct MaxNumStepsOutput