+++ CyRK/cy/cysolve.cpp
@@ -16,6 +16,7 @@ std::shared_ptr<CySolverResult> baseline_cysolve_ivp(
     const bool dense_output,
     const double* t_eval,
     const size_t len_t_eval,
+    PreEvalFunc pre_eval_func,
     // rk optional arguments
     const double rtol,
     const double atol,
@@ -85,7 +86,7 @@ std::shared_ptr<CySolverResult> baseline_cysolve_ivp(
         solver = new RK23(
             // Common Inputs
             diffeq_ptr, solution_ptr, t_start, t_end, y0_ptr, num_y, num_extra, args_ptr, max_num_steps, max_ram_MB,
-            dense_output, t_eval, len_t_eval,
+            dense_output, t_eval, len_t_eval, pre_eval_func,
             // RK Inputs
             rtol, atol, rtols_ptr, atols_ptr, max_step_size, first_step_size
         );
@@ -95,7 +96,7 @@ std::shared_ptr<CySolverResult> baseline_cysolve_ivp(
         solver = new RK45(
             // Common Inputs
             diffeq_ptr, solution_ptr, t_start, t_end, y0_ptr, num_y, num_extra, args_ptr, max_num_steps, max_ram_MB,
-            dense_output, t_eval, len_t_eval,
+            dense_output, t_eval, len_t_eval, pre_eval_func,
             // RK Inputs
             rtol, atol, rtols_ptr, atols_ptr, max_step_size, first_step_size
         );
@@ -105,7 +106,7 @@ std::shared_ptr<CySolverResult> baseline_cysolve_ivp(
         solver = new DOP853(
             // Common Inputs
             diffeq_ptr, solution_ptr, t_start, t_end, y0_ptr, num_y, num_extra, args_ptr, max_num_steps, max_ram_MB,
-            dense_output, t_eval, len_t_eval,
+            dense_output, t_eval, len_t_eval, pre_eval_func,
             // RK Inputs
             rtol, atol, rtols_ptr, atols_ptr, max_step_size, first_step_size
         );
@@ -181,6 +182,9 @@ PySolver::PySolver(
     // We need to pass a fake diffeq pointer (diffeq ptr is unused in python-based solver)
     DiffeqFuncType diffeq_ptr = nullptr;
 
+    // We also need to pass a fake pre-eval function
+    PreEvalFunc pre_eval_func = nullptr;
+
     // Build the solver class. This must be heap allocated to take advantage of polymorphism.
     switch (this->integration_method)
     {
@@ -201,6 +205,7 @@ PySolver::PySolver(
             dense_output,
             t_eval,
             len_t_eval,
+            pre_eval_func,
             // rk optional arguments
             rtol,
             atol,
@@ -226,6 +231,7 @@ PySolver::PySolver(
             dense_output,
             t_eval,
             len_t_eval,
+            pre_eval_func,
             // rk optional arguments
             rtol,
             atol,
@@ -251,6 +257,7 @@ PySolver::PySolver(
             dense_output,
             t_eval,
             len_t_eval,
+            pre_eval_func,
             // rk optional arguments
             rtol,
             atol,