+++ CyRK/cy/cysolve.hpp
@@ -22,6 +22,7 @@ std::shared_ptr<CySolverResult> baseline_cysolve_ivp(
     const bool dense_output = false,
     const double* t_eval = nullptr,
     const size_t len_t_eval = 0,
+    PreEvalFunc pre_eval_func = nullptr,
     // rk optional arguments
     const double rtol = 1.0e-3,
     const double atol = 1.0e-6,